% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_taxa.R
\name{aggregate_taxa}
\alias{aggregate_taxa}
\title{Aggregates a MRexperiment object and returns either an aggregated
MRexperiment or counts matrix to a user defined taxonomic level.}
\usage{
aggregate_taxa(obj, lvl, aggfun = colSums, out = "MRexperiment", ...)
}
\arguments{
\item{obj}{A MRexperiment-class object or count matrix.}

\item{lvl}{featureData column name from the MRexperiment object or if count
matrix object a vector of labels.}

\item{aggfun}{Matrix aggregation function, e.g. colSums.}

\item{out}{Either 'MRexperiment' or 'matrix'}

\item{...}{Additional parameters to pass to MRcount, e.g. norm, log, and sl.}
}
\value{
An aggregated count matrix or MRexperiment
}
\description{
Using the featureData information in the MRexperiment-class object,
aggregate_taxa aggregates the OTU count data (MRexperiment assayData slot)
to a user defined taxonomic level (i.e. 'genus') using the defined
aggfun function (default colSums). Possible aggfun alternatives include and
column wise matrix calculations, e.g. colMeans, colMedians.
}
\examples{
data("mouseData", package = "metagenomeSeq")
aggregate_taxa(mouseData[1:100,],lvl="class",norm=TRUE,aggfun=colSums)
aggregate_taxa(mouseData,lvl="class",norm=TRUE,aggfun=colSums)
aggregate_taxa(mouseData,lvl='phylum',norm=FALSE,aggfun=colSums)
}
