% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb_aggregate_taxa.R
\docType{methods}
\name{aggregate_taxonomy}
\alias{aggregate_taxonomy}
\alias{aggregate_taxonomy,MgDb-method}
\title{Taxonomy aggregation for a given keytype}
\usage{
aggregate_taxonomy(mgdb, taxa_level, mapping)

\S4method{aggregate_taxonomy}{MgDb}(mgdb, taxa_level, mapping)
}
\arguments{
\item{mgdb}{object of MgDB class}

\item{taxa_level}{taxonomic aggregation level}

\item{mapping}{logical for returning mapping of ids to aggregated taxa level}
}
\value{
list(tbl_df, tbl_df)
}
\description{
Taxonomy aggregation for a given keytype
}
\examples{
gg97 <- get_gg13.8_97MgDb()
aggregate_taxonomy(gg97, taxa_level = "Phylum")
}
