% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_taxa.R
\docType{methods}
\name{aggregate_taxonomy_mgf}
\alias{aggregate_taxonomy_mgf}
\alias{aggregate_taxonomy_mgf,mgFeatures-method}
\alias{aggregate_taxonomy_mgf}
\title{Taxonomy aggregation for a given level}
\usage{
aggregate_taxonomy_mgf(mgf, ...)

\S4method{aggregate_taxonomy_mgf}{mgFeatures}(mgf, ...)
}
\arguments{
\item{mgf}{object of mgFeatures class}

\item{taxa_level}{taxonomic aggregation level}

\item{mapping}{logical for returning mapping of ids to aggregated taxa level}
}
\value{
list(tbl_df, tbl_df)
}
\description{
Taxonomy aggregation for a given level
}
\examples{
data(mock_mgF)
aggregate_taxonomy_mgf(mock_mgF, taxa_level = "Phylum")
}
