% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_taxa.R
\name{aggregateByTaxonomy}
\alias{aggTax}
\alias{aggregateByTaxonomy}
\title{Aggregates a MRexperiment object or counts matrix to a user defined
  taxonomic level.}
\usage{
aggregateByTaxonomy(obj, lvl, norm = FALSE, log = FALSE, aggfun = colSums,
  sl = 1000, out = "MRexperiment")

aggTax(obj, lvl, norm = FALSE, log = FALSE, aggfun = colSums, sl = 1000,
  out = "MRexperiment")
}
\arguments{
\item{obj}{A MRexperiment-class object or count matrix.}

\item{lvl}{featureData column name from the MRexperiment object or if count
matrix object a vector of labels.}

\item{norm}{Whether to aggregate normalized counts or not.}

\item{log}{Whether or not to log2 transform the counts - if MRexperiment
object.}

\item{aggfun}{Matrix aggregation function, e.g. colSums.}

\item{sl}{scaling value, default is 1000.}

\item{out}{Either 'MRexperiment' or 'matrix'}
}
\value{
An aggregated count matrix or MRexperiment
}
\details{
Using the featureData information in the MRexperiment-class object, calling
  aggregateByTaxonomy on a MRexperiment and a particular featureData column
  (i.e. 'genus') will aggregate counts to the desired level using the aggfun
  function (default colSums). Possible aggfun alternatives include and column
  wise matrix calculations, e.g. colMeans, colMedians.
}
\examples{
# not run
# data("mouseData", package = "metagenomeSeq")
# aggregateByTaxonomy(mouseData[1:100,],lvl="class",norm=TRUE,aggfun=colSums)
# aggregateByTaxonomy(mouseData,lvl="class",norm=TRUE,aggfun=colMedians)
# aggTax(mouseData,lvl='phylum',norm=FALSE,aggfun=colSums)
}

