% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb-class.R
\docType{methods}
\name{annotateMRexp,MgDb-method}
\alias{annotateMRexp,MgDb-method}
\title{Annotate MRexperiment object with seq taxonomy from MgDb object}
\usage{
\S4method{annotateMRexp}{MgDb}(mgdb, MRobj)
}
\arguments{
\item{mgdb}{MgDb class object}

\item{MRobj}{MRexperiment class object}

\item{...}{additional arguments passed to select function}
}
\value{
metagenomeAnnotation-class object
}
\description{
This method is used annotate a MRexperiment with taxonomic information from a \link[=MgDb]{MgDb-class}
object using the MRexperiment object's Feature names.
object.
}
\examples{
# see vignette
}

