% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb_method_annotateMRexp.R
\docType{methods}
\name{annotateMRexp_fData}
\alias{annotateMRexp_fData}
\alias{annotateMRexp_fData,MgDb-method}
\title{Annotate MRexperiment object featureData slot using MgDb object}
\usage{
annotateMRexp_fData(mgdb, MRobj, ...)

\S4method{annotateMRexp_fData}{MgDb}(mgdb, MRobj)
}
\arguments{
\item{mgdb}{MgDb class object}

\item{MRobj}{MRexperiment class object}

\item{...}{additional arguments passed to select function}
}
\value{
MRexperiment-class object
}
\description{
This method is used to define a MRexperiment object featureData slot with
taxonomic information from a \link[=MgDb]{MgDb-class} object using the
MRexperiment object's Feature names. object.
}

