% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb-class.R
\name{.load_tree}
\alias{.load_tree}
\title{Tree can be either rds with phylo class object or tree file}
\usage{
.load_tree(tree_file)
}
\arguments{
\item{tree_file}{location of tree}
}
\description{
Tree can be either rds with phylo class object or tree file
}
\keyword{internals}
