% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb_method_annotateFeatures.R
\name{.mgDb_annotateFeatures}
\alias{.mgDb_annotateFeatures}
\title{Generates an mgFeatures object when passed a vector of keys or dataframe
(e.g. query)}
\usage{
.mgDb_annotateFeatures(mgdb, query)
}
\arguments{
\item{mgdb}{mdgdb object}

\item{query}{search param}
}
\description{
Generates an mgFeatures object when passed a vector of keys or dataframe
(e.g. query)
}
\keyword{internals}
