% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb_method_select.R
\name{.select}
\alias{.select}
\title{either select by ids for taxa information}
\usage{
.select(mgdb, type, keys, keytype, columns)
}
\arguments{
\item{mgdb}{mgdb object}

\item{type}{selection type, can be seq, taxa, tree or all}

\item{keys}{keys to filter}

\item{keytype}{keytypes to select}

\item{columns}{columns to select, defaults to all}
}
\description{
either select by ids for taxa information
}
\keyword{internals}
