% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb_method_select.R
\name{.select.seq}
\alias{.select.seq}
\title{select sequences}
\usage{
.select.seq(seqObj, row_names, Keys)
}
\arguments{
\item{seqObj}{seq object}

\item{row_names}{row names to select}

\item{Keys}{keys to filter}
}
\description{
select sequences
}
\keyword{internals}
