% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb_method_select.R
\name{.select.taxa}
\alias{.select.taxa}
\title{select taxa}
\usage{
.select.taxa(taxaDb, metaDb, keys, keytype, columns = "all")
}
\arguments{
\item{taxaDb}{taxa object}

\item{metaDb}{metadata object}

\item{keys}{keys to filter}

\item{keytype}{keytypes to select}

\item{columns}{columns to select, defaults to all}
}
\description{
select taxa
}
\keyword{internals}
