% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb_method_select.R
\name{.select.tree}
\alias{.select.tree}
\title{select tree tips}
\usage{
.select.tree(tree, ids)
}
\arguments{
\item{tree}{tree object}

\item{ids}{tip ids to select}
}
\description{
select tree tips
}
\keyword{internals}
