% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb-taxa_accessors.R
\name{.taxa_keys}
\alias{.taxa_keys}
\title{Taxa keys functions
import dplyr}
\usage{
.taxa_keys(mgdb, keytype)
}
\arguments{
\item{mgdb}{mgDb class}

\item{keytype}{taxa keytype}
}
\description{
Taxa keys functions
import dplyr
}
\keyword{internals}
