% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoMgDb.R
\name{get_demoMgDb}
\alias{get_demoMgDb}
\title{Example MgDb-class object}
\usage{
get_demoMgDb()
}
\value{
MgDb-class object
}
\description{
Example \link[=MgDb]{MgDb-class} object with 3211 entries from the Greengenes
13.8 OTU 99 database.
}
\examples{
 get_demoMgDb()
}
