% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg13.8_97MgDb.R
\name{get_gg13.8_97MgDb}
\alias{get_gg13.8_97MgDb}
\title{MgDb-class object with Greengenes Database Version 13.8 97\% OTUs}
\usage{
get_gg13.8_97MgDb()
}
\value{
MgDb-class object
}
\description{
Example \link[=MgDb]{MgDb-class} object with Greengenes Database Version
13.8 97\% OTUs.
}
\examples{
get_gg13.8_97MgDb()
}
