% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb-class.R
\name{make_mgdb}
\alias{make_mgdb}
\title{make_mgdb}
\usage{
make_mgdb(db_name, db_file, taxa_tbl, seqs)
}
\arguments{
\item{db_name}{reference database name}

\item{db_file}{file path for sqlite database}

\item{taxa_tbl}{data frame with database taxonomy data}

\item{seqs}{database sequences, path to fasta file or DNAStringSet object}
}
\description{
make_mgdb
}
