% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeMgDbFromGreenGenes.R
\name{makeMgDbFromGreenGenes}
\alias{makeMgDbFromGreenGenes}
\title{makeMgDb for GreenGenes database}
\usage{
makeMgDbFromGreenGenes(db_name = "gg_13_5",
  taxa_url = "ftp://greengenes.microbio.me/greengenes_release/gg_13_5/gg_13_5_taxonomy.txt.gz",
  seq_url = "ftp://greengenes.microbio.me/greengenes_release/gg_13_5/gg_13_5.fasta.gz",
  metadata = list(URL = "https://greengenes.microbio.me", DB_TYPE_NAME =
  "GreenGenes"))
}
\arguments{
\item{db_name}{}

\item{taxa_url}{}

\item{seq_url}{}

\item{metadata}{}
}
\value{
mgDb object
}
\description{
makeMgDb for GreenGenes database
}
\examples{
makeMgDbFromGreenGenes()
}

