% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mgDb-class.R
\docType{class}
\name{MgDb}
\alias{MgDb}
\title{Metagenome Database class}
\description{
Metagenome Database class
}
\section{Fields}{

\describe{
\item{\code{taxa}}{taxonomy database}

\item{\code{seq}}{database reference sequences}

\item{\code{metadata}}{database metadata}
}}

