% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb-class.R
\name{mgDb_}
\alias{mgDb_}
\alias{mgDb_tree}
\alias{mgDb_seq}
\alias{mgDb_taxa}
\alias{mgDb_meta}
\title{MgDb-class accessors}
\usage{
mgDb_tree(mgdb)

mgDb_seq(mgdb)

mgDb_taxa(mgdb)

mgDb_meta(mgdb)
}
\arguments{
\item{mgdb}{MgDb-class object.}
}
\value{
appropriate class object for the slot accessed
}
\description{
Accessors for \linkS4class{MgDb}-class object slots. \code{mgDb_seq} -
sequence slot, \code{mgDb_taxa} - taxa slot, \code{mgDb_tree} - phylogenetic
tree slot, and \code{mgDb_meta} - metadata slot.
}
\examples{
gg85 <- get_gg13.8_85MgDb()
mgDb_seq(gg85)
mgDb_taxa(gg85)
mgDb_tree(gg85)
mgDb_meta(gg85)
}
