% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb-taxa_accessors.R
\docType{methods}
\name{taxa_keys}
\alias{taxa_columns}
\alias{taxa_columns,ANY-method}
\alias{taxa_columns,MgDb-method}
\alias{taxa_keys}
\alias{taxa_keys,ANY-method}
\alias{taxa_keys,MgDb-method}
\alias{taxa_keytypes,ANY-method}
\alias{taxa_keytypes,MgDb-method}
\title{Taxonomy values for a given keytype}
\usage{
taxa_keys(mgdb, keytype)

\S4method{taxa_keys}{MgDb}(mgdb, keytype)

taxa_columns(mgdb)

\S4method{taxa_columns}{MgDb}(mgdb)

taxa_keytypes(mgdb)

\S4method{taxa_keytypes}{MgDb}(mgdb)
}
\arguments{
\item{mgdb}{object of MgDB class}

\item{keytype}{taxonomic classification level}

\item{mgdb}{object of MgDB class}

\item{mgdb}{object of MgDB class}
}
\value{
tbl_df

tbl_df
}
\description{
Taxonomy values for a given keytype

Column names for MgDb taxonomy slot object

Column names for MgDb taxonomy slot object
}
\note{
Same function as \code{\link{taxa_keytypes}}.
}
\examples{
demoMgDb <- get_demoMgDb()
taxa_keys(demoMgDb, keytype = "Phylum")
demoMgDb <- get_demoMgDb()
taxa_columns(demoMgDb)
demoMgDb <- get_demoMgDb()
taxa_keytypes(demoMgDb)
}

