% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_taxa.R
\docType{methods}
\name{mgDb_aggregate_taxonomy}
\alias{mgDb_aggregate_taxonomy}
\alias{mgDb_aggregate_taxonomy,MgDb-method}
\title{Aggregate MgDb class object to a taxonomic level}
\usage{
mgDb_aggregate_taxonomy(mgdb, ...)

\S4method{mgDb_aggregate_taxonomy}{MgDb}(mgdb, ...)
}
\arguments{
\item{mgdb}{object of MgDB class}

\item{taxa_level}{taxonomic aggregation level}

\item{mapping}{logical for returning mapping of ids to aggregated taxa level}
}
\value{
tbl_df or list(tbl_df, tbl_df)
}
\description{
Returns a tibble with distinct rows containing the taxonomic heirarchy to a
user defined taxonomic level. Optionally a mapping tibble for use in
aggregating counts, with two columns one with values for user defined
taxonomic level and sequence keys.
}
\examples{
gg97 <- get_gg13.8_97MgDb()
mgDb_aggregate_taxonomy(gg97, taxa_level = "Phylum")
}
