% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb-class.R
\name{mgdb_seq}
\alias{mgdb_seq}
\title{MgDb seq slot accessor}
\usage{
mgdb_seq(mgdb)
}
\arguments{
\item{mgdb}{MgDb class object}
}
\value{
DNAStringSet class object
}
\description{
MgDb seq slot accessor
}
\examples{
gg85 <- get_gg13.8_85MgDb()
mgdb_seq(gg85)
}
