% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb-class.R
\name{mgdb_taxa}
\alias{mgdb_taxa}
\title{MgDb taxa slot accessor}
\usage{
mgdb_taxa(mgdb)
}
\arguments{
\item{mgdb}{MgDb class object}
}
\value{
tbl_sql connection to sqlite table
}
\description{
MgDb taxa slot accessor
}
\examples{
demoMgDb <- get_demoMgDb()
mgdb_taxa(demoMgDb)
}

