% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb-class.R
\name{mgdb_tree}
\alias{mgdb_tree}
\title{MgDb tree slot accessor}
\usage{
mgdb_tree(mgdb)
}
\arguments{
\item{mgdb}{MgDb class object}
}
\value{
phylo class object
}
\description{
MgDb tree slot accessor
}
\examples{
gg97 <- get_gg13.8_97MgDb()
mgdb_tree(gg97)
}
