% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgFeatures-class.R
\name{mgF_}
\alias{mgF_}
\alias{mgF_tree}
\alias{mgF_seq}
\alias{mgF_taxa}
\alias{mgF_meta}
\title{mgFeatures accessors}
\usage{
mgF_tree(mgF)

mgF_seq(mgF)

mgF_taxa(mgF)

mgF_meta(mgF)
}
\arguments{
\item{mgF}{mgFeatures-class object.}
}
\value{
appropriate class object for the slot accessed
}
\description{
Accessors for \linkS4class{mgFeatures}-class object slots. \code{mgF_seq} - refDbSeq slot, \code{mgF_taxa} - taxa slot, \code{mgF_tree} - phylogenetic tree slot, and \code{mgF_meta} - metadata slot.
}
\examples{
data(mock_mgF)
mgF_seq(mock_mgF)
mgF_taxa(mock_mgF)
mgF_tree(mock_mgF)
mgF_meta(mock_mgF)
}
