% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_taxa.R
\docType{methods}
\name{mgF_aggregate_taxonomy}
\alias{mgF_aggregate_taxonomy}
\alias{mgF_aggregate_taxonomy,mgFeatures-method}
\alias{mgF_aggregate_taxonomy}
\title{Aggregate mgFeatures class object to a taxonomic level}
\usage{
mgF_aggregate_taxonomy(mgf, ...)

\S4method{mgF_aggregate_taxonomy}{mgFeatures}(mgf, ...)
}
\arguments{
\item{mgf}{object of mgFeatures class}

\item{taxa_level}{taxonomic aggregation level}

\item{mapping}{logical for returning mapping of ids to aggregated taxa level}
}
\value{
list(tbl_df, tbl_df)
}
\description{
Returns a tibble with distinct rows containing the taxonomic heirarchy to a
user defined taxonomic level. Optionally a mapping tibble for use in
aggregating counts, with two columns one with values for user defined
taxonomic level and sequence keys.
}
\examples{
data(mock_mgF)
mgF_aggregate_taxonomy_mgf(mock_mgF, taxa_level = "Phylum")
}
