% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgFeatures-class.R
\name{mgF_meta}
\alias{mgF_meta}
\title{mgFeatures metadata slot accessor}
\usage{
mgF_meta(mgF)
}
\arguments{
\item{mgF}{mgFeatures class object}
}
\value{
list
}
\description{
mgFeatures metadata slot accessor
}
\examples{
data(mock_mgF)
mgF_meta(mock_mgF)
}

