% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgFeatures-class.R
\name{mgF_seq}
\alias{mgF_seq}
\title{mgFeatures referenceDbSeqData slot accessor}
\usage{
mgF_seq(mgF)
}
\arguments{
\item{mgF}{mgFeatures class object}
}
\value{
DNAStringSet class object
}
\description{
mgFeatures referenceDbSeqData slot accessor
}
\examples{
## mgF_tree(demo_mgF)
}

