% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgFeatures-class.R
\name{mgF_taxa}
\alias{mgF_taxa}
\title{mgFeatures tree slot accessor}
\usage{
mgF_taxa(mgF)
}
\arguments{
\item{mgF}{mgFeatures class object}
}
\value{
AnnotatedDataFrame
}
\description{
mgFeatures tree slot accessor
}
\examples{
## mgF_tree(demo_mgF)
}

