% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgFeatures-class.R
\name{mgF_tree}
\alias{mgF_tree}
\title{mgFeatures referenceDbTreeData slot accessor}
\usage{
mgF_tree(mgF)
}
\arguments{
\item{mgF}{mgFeatures class object}
}
\value{
phylo class object
}
\description{
mgFeatures referenceDbTreeData slot accessor
}
\examples{
## mgF_tree(demo_mgF)
}

