% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgFeatures-class.R
\docType{methods}
\name{mgFeatures-methods}
\alias{mgFeatures-methods}
\alias{[,mgFeatures-method}
\title{mgFeatures-class subset method}
\usage{
\S4method{[}{mgFeatures}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{Object to extract elements}

\item{i, j}{element indices to extract or replace}

\item{...}{other parameters to subset function}

\item{drop}{default to FALSE}
}
\description{
mgFeatures-class subset method
}
