% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgFeatures-class.R
\name{mgFeatures}
\alias{mgFeatures}
\title{mgFeatures-class constructor}
\usage{
mgFeatures(taxa, tree, seq, metadata)
}
\arguments{
\item{taxa}{a DataFrame-class or object that can be coerced into a DataFrame}

\item{tree}{a phylo-class object with phylogenetic tree}

\item{seq}{DNAStringSet-object with feature sequences}

\item{metadata}{a list}
}
\description{
mgFeatures-class constructor
}
