% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgQuery.R
\docType{data}
\name{mgQuery}
\alias{mgQuery}
\title{Example ShortRead 16S experiment dataset}
\format{A metagenomeFeatures object with 500 sequences.}
\source{
\url{http://bioconductor.org/packages/release/data/experiment/html/msd16s.html}
}
\usage{
mgQuery
}
\description{
A dataset containing the top OTU centers from a 16S marker gene survey.
Specifically, the OTU centers came from 'Healthy and moderate to severe
diarrhea 16S expression data' - bioconductor dataset 'msd16s'.
}
\examples{
data(mgQuery)
}
\keyword{datasets}
