% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_mgF.R
\docType{data}
\name{mock_mgF}
\alias{mock_mgF}
\title{Example mgFeatures class object}
\format{mgFeatures}
\usage{
mock_mgF
}
\description{
Example \link{mgFeatures-class} object generated using the
\link{annotateFeatures} \link[=MgDb]{MgDb-class} method. The dataset contains
the OTU ids and Greengenes database version 13.5 ids. Specifically, the OTU
centers came from a mock community dataset from the study Bokulich et al.
2013, the OTU ids were extracted from a biom file downloaded from QIITA
(https://qiita.ucsd.edu).
}
\details{
Bokulich, Nicholas A., et al. "Quality-filtering vastly improves diversity
estimates from Illumina amplicon sequencing." Nature methods 10.1 (2013): 57-59.
}
\examples{
data(mock_mgF)
}
\keyword{datasets}

