% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/msd16s_metagenomeAnnotation.R
\docType{data}
\name{msd16s_metagenomeAnnotation}
\alias{msd16s_metagenomeAnnotation}
\title{Example metagenomeAnnotation object}
\format{A metagenomeAnnotation class object.}
\source{
\url{http://bioconductor.org/packages/release/data/experiment/html/msd16s.html}
}
\usage{
msd16s_metagenomeAnnotation
}
\description{
This example \link{metagenomeAnnotation-class} object
includes \link[msd16s]{msd16s} OTU centers from a 16S marker gene
survey. Specifically, the OTU centers came from 'Healthy and moderate to
severe diarrhea 16S expression data' - bioconductor dataset 'msd16s'.
}
\examples{
data(msd16s_metagenomeAnnotation)
}
\keyword{datasets}

