% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/msd16s_query_df.R
\docType{data}
\name{msd16s_query_df}
\alias{msd16s_query_df}
\title{Example Query Data Frame for creating metagenomeAnnotation object}
\format{data.frame}
\source{
\url{http://bioconductor.org/packages/release/data/experiment/html/msd16s.html}
}
\usage{
msd16s_query_df
}
\description{
A dataset containing the msd16s OTU ids and Greengenes database version 13.5 ids.
Specifically, the OTU centers came from 'Healthy and moderate to severe
diarrhea 16S expression data' - bioconductor dataset 'msd16s'.
}
\keyword{datasets}

