% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qiita_study_94_gg_ids.R
\docType{data}
\name{qiita_study_94_gg_ids}
\alias{qiita_study_94_gg_ids}
\title{QIITA Dataset
For this vignette we are using 16S rRNA data from Rousk et al. 2010, a soil microbiome study, https://qiita.ucsd.edu/study/description/94.
A BIOM and qiime mapping file for the study can be obtained from QIITA.
A vector of Greengenes for the study cluster centers is included in this package for use in this vignette.}
\format{An object of class \code{character} of length 2305.}
\usage{
qiita_study_94_gg_ids
}
\description{
data(qiita_study_94_gg_ids)
}
\keyword{datasets}
