% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mgDb-class.R
\docType{methods}
\name{select}
\alias{select}
\alias{select,MgDb-method}
\title{Function for querying MgDb class objects
 Use keys - keytype sets to define subset.
 keytype - is the taxonomic level with the taxa of interest, use "Keys" if selecting based on selecting based on sequence IDs.
 keys - is a vector of values of keytype}
\usage{
select(mgdb, type, ...)

\S4method{select}{MgDb}(mgdb, type, keys = NULL, keytype = NULL,
  columns = "all")
}
\arguments{
\item{mgdb}{MgDb class object}

\item{type}{either "taxa", "seq", or "both". "taxa" and "seq" only queries the taxonomy and sequences databases respectively. "both" queries both the taxonomy and sequence database.}

\item{...}{additional arguments passed to select function}

\item{keys}{specific taxonomic groups to select for}

\item{keytype}{taxonomic level of keys}

\item{columns}{keytypes in taxonomy databse to return, all by default}
}
\value{
generates database, function does not return anything
}
\description{
Function for querying MgDb class objects
 Use keys - keytype sets to define subset.
 keytype - is the taxonomic level with the taxa of interest, use "Keys" if selecting based on selecting based on sequence IDs.
 keys - is a vector of values of keytype
}
\examples{
library(greengenes13.5MgDb)
# select taxa only
select(gg13.5MgDb, type = "taxa",
     keys = c("Vibrio", "Salmonella"),
     keytype = "Genus")

 # select seq only
select(gg13.5MgDb, type = "seq",
      keys = c("Vibrio", "Salmonella"),
      keytype = "Genus")

# select both taxa and seq
select(gg13.5MgDb, type = "both",
       keys = c("Vibrio", "Salmonella"),
       keytype = "Genus")
}

