% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metagenomeAnnotation-class.R
\docType{methods}
\name{show,metagenomeAnnotation-method}
\alias{show,metagenomeAnnotation-method}
\title{Display summary of metagenomeAnnotaiton-class object}
\usage{
\S4method{show}{metagenomeAnnotation}(object)
}
\arguments{
\item{object}{metagenomeAnnotation-class object}
}
\description{
Display summary of metagenomeAnnotaiton-class object
}

