% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metagenomeAnnotation-class.R
\docType{methods}
\name{split_by,metagenomeAnnotation-method}
\alias{split_by,metagenomeAnnotation-method}
\title{Split metagenomeAnnotation Object by Taxa}
\usage{
\S4method{split_by}{metagenomeAnnotation}(mgAnno, taxa_level)
}
\arguments{
\item{mgAnno}{metagenomeAnnotation class object}

\item{taxa_level}{taxonomic level used to split the metagenomeAnnotation object at}
}
\value{
list of metagenomeAnnotation objects
}
\description{
Split metagenomeAnnotation Object by Taxa
}

