% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb-taxa_accessors.R
\docType{methods}
\name{taxa_}
\alias{taxa_}
\alias{taxa_keys}
\alias{taxa_keys,MgDb-method}
\alias{taxa_columns}
\alias{taxa_columns,MgDb-method}
\alias{taxa_keytypes}
\alias{taxa_keytypes,MgDb-method}
\title{MgDb-class Taxa slot helper functions}
\usage{
taxa_keys(mgdb, keytype)

\S4method{taxa_keys}{MgDb}(mgdb, keytype)

taxa_columns(mgdb)

\S4method{taxa_columns}{MgDb}(mgdb)

taxa_keytypes(mgdb)

\S4method{taxa_keytypes}{MgDb}(mgdb)
}
\arguments{
\item{mgdb}{MgDb-class object.}

\item{keytype}{character string specifying keys to return}
}
\value{
appropriate class object for the slot accessed
}
\description{
Helper functions for for \linkS4class{MgDb}-class taxa slot.
\code{taxa_columns} - taxa slot column names, \code{taxa_keytypes} - taxa
slot keytypes (values used with \code{taxa_columns} and \code{mgDb_select}
functions), and \code{taxa_keys} - database values for a specific keytype.
}
\examples{
gg85 <- get_gg13.8_85MgDb()
taxa_columns(gg85)
taxa_keytypes(gg85)
taxa_keys(gg85, keytype = "Phylum")
}
