% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mgDb-taxa_accessors.R
\docType{methods}
\name{taxa_columns}
\alias{taxa_columns}
\alias{taxa_columns,MgDb-method}
\title{Column names for MgDb taxonomy slot object}
\usage{
taxa_columns(mgdb)

\S4method{taxa_columns}{MgDb}(mgdb)
}
\arguments{
\item{mgdb}{object of MgDB class}
}
\value{
tbl_df
}
\description{
Column names for MgDb taxonomy slot object
}

