% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mgDb-class.R
\name{taxa_columns}
\alias{taxa_columns}
\title{Taxonomic levels for a MdDB class}
\usage{
taxa_columns(mgdb_object)
}
\arguments{
\item{mgdb_object}{object of MgDB class}

\item{keytype}{taxonomic classification level}
}
\value{
vector
}
\description{
Taxonomic levels for a MdDB class
}
\examples{
taxa_keys(mgdb)
}

