% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mgDb-taxa_accessors.R
\docType{methods}
\name{taxa_keys,MgDb-method}
\alias{taxa_keys,MgDb-method}
\title{Taxonomy values for a given keytype}
\usage{
\S4method{taxa_keys}{MgDb}(x, ...)
}
\arguments{
\item{mgdb}{object of MgDB class}

\item{keytype}{taxonomic classification level}
}
\value{
tbl_df
}
\description{
Taxonomy values for a given keytype
}
\examples{
taxa_keys(mgdb, keytype = "Class")
}

