% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mgDb-class.R
\name{taxa_keys}
\alias{taxa_keys}
\title{Taxonomy values for a given keytype}
\usage{
taxa_keys(mgdb_object, keytype)
}
\arguments{
\item{mgdb_object}{object of MgDB class}

\item{keytype}{taxonomic classification level}
}
\value{
tbl_df
}
\description{
Taxonomy values for a given keytype
}
\examples{
taxa_keys(mgdb, "Class")
}

