% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mgDb-class.R
\name{taxa_keytypes}
\alias{taxa_keytypes}
\title{Keytypes for the MgDB object}
\usage{
taxa_keytypes(mgdb_object)
}
\arguments{
\item{mgdb_object}{object of MgDB class}
}
\value{
vector
}
\description{
Keytypes for the MgDB object
}
\examples{
taxa_keytypes(mgdb)
}

