% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_taxa.R
\name{taxa_levels}
\alias{taxa_levels}
\title{Accessior function for taxanomic levels in MRexperiment featureData}
\usage{
taxa_levels(obj)
}
\arguments{
\item{obj}{an `MRexperiment-class` object}
}
\value{
character vector with taxonomic levels
}
\description{
Accessior function for taxanomic levels in MRexperiment featureData
}
\examples{
# not run
# data("mouseData", package = "metagenomeSeq")
# taxa_levels(mouseData)
}

