% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mgDb-class.R
\docType{methods}
\name{annotate,MgDb-method}
\alias{annotate,MgDb-method}
\title{annotating a set of sequences with taxonomic information from a MgDb class object}
\usage{
\S4method{annotate}{MgDb}(x, query, mapping, ...)
}
\arguments{
\item{x}{MgDb class object}

\item{query}{ShortRead-class object with marker gene sequences}

\item{mapping}{method used to map sequences to database}
}
\value{
metagenomeAnnotation class object
}
\description{
annotating a set of sequences with taxonomic information from a MgDb class object
}

