% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb_method_annotateFeatures.R
\docType{methods}
\name{annotateFeatures}
\alias{annotate,MgDb-method}
\alias{annotateFeatures}
\alias{annotateFeatures,MgDb-method}
\title{Annotating metagenome data with taxonomic information}
\usage{
annotateFeatures(mgdb, ...)

\S4method{annotateFeatures}{MgDb}(mgdb, query_key)
}
\arguments{
\item{mgdb}{MgDb class object}

\item{...}{additional arguments passed to select function}

\item{db_keys}{(Optional) vector of database Keys of entries to include in
metagenomeAnnotation class object}

\item{query_df}{(Optional) data frame with experimental data to annotate with
taxonomic information, must include column named "Key" with databse ids.}
}
\value{
mgFeatures-class object
}
\description{
This method is used to create a \linkS4class{mgFeatures} class
object
}
\note{
Must include either db_keys or query_df as argument.
}
\examples{
# see vignette
}

