% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockMgDb.R
\name{get_mockMgDb}
\alias{get_mockMgDb}
\title{Mock MgDb-class object}
\usage{
get_mockMgDb()
}
\value{
MgDb-class object
}
\description{
Mock \link[=MgDb]{MgDb-class} object from the Greengenes 13.5 database.
}
\examples{
get_mockMgDb()
}

