% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb-class.R
\name{make_mgdb_sqlite}
\alias{make_mgdb_sqlite}
\title{make_mgdb_sqlite}
\usage{
make_mgdb_sqlite(db_name, db_file, taxa_tbl, seqs)
}
\arguments{
\item{db_name}{reference database name}

\item{db_file}{file path for sqlite database}

\item{taxa_tbl}{data frame with database taxonomy data}

\item{seqs}{database sequences, path to fasta file or DNAStringSet object}
}
\value{
writes SQLite file
}
\description{
make_mgdb_sqlite
}
\examples{
\dontrun{
make_mgdb_sqlite(db_name = "greengenes13.8_97",
    db_file = db_file,
    taxa_tbl = taxa_tbl,
    seqs = seqs)
}
}
\keyword{internal}
