% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metagenomeAnnotation-class.R
\docType{class}
\name{metagenomeAnnotation-class}
\alias{metagenomeAnnotation-class}
\alias{mgAnno}
\alias{show,metagenomeAnnotation-method}
\title{annotated marker gene sequence object class}
\usage{
\S4method{show}{metagenomeAnnotation}(object)
}
\arguments{
\item{object}{metagenomeAnnotation-class object}
}
\value{
metagenomeAnnotation class object
}
\description{
annotated marker gene sequence object class

Display summary of metagenomeAnnotaiton-class object
}
\section{Slots}{

\describe{
\item{\code{annotationData}}{AnnotatedDataFrame}

\item{\code{metadata}}{list}

\item{\code{featureData}}{DNAStringSet}
}}

