% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metagenomeAnnotation-class.R
\docType{class}
\name{metagenomeAnnotation-class}
\alias{metagenomeAnnotation-class}
\alias{mgAnno}
\title{annotated marker gene sequence object class}
\value{
metagenomeAnnotation class object
}
\description{
annotated marker gene sequence object class
}
\section{Slots}{

\describe{
\item{\code{annotationData}}{AnnotatedDataFrame}

\item{\code{metadata}}{list}

\item{\code{featureData}}{DNAStringSet}
}}

