% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metagenomeAnnotation-class.R
\docType{class}
\name{metagenomeAnnotation-class}
\alias{metagenomeAnnotation-class}
\title{metagenomeAnnotation-class object
Object contains marker gene sequence object with taxonomic annotation data
for set of sequences.  The class extends the AnnotatedDataFrame class with
a slot for metadata including information on the database source and methods
used to perform the taxonomic assignment. Additionally, the user can include
cluster sequence experiment sequence data in the object for use in downstream
analysis.}
\value{
metagenomeAnnotation class object
}
\description{
metagenomeAnnotation-class object
Object contains marker gene sequence object with taxonomic annotation data
for set of sequences.  The class extends the AnnotatedDataFrame class with
a slot for metadata including information on the database source and methods
used to perform the taxonomic assignment. Additionally, the user can include
cluster sequence experiment sequence data in the object for use in downstream
analysis.
}
\section{Slots}{

\describe{
\item{\code{metadata}}{list}

\item{\code{experimentSeqData}}{DNAStringSet}
}}

