% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomeAnnotation-class.R
\docType{class}
\name{metagenomeAnnotation-class}
\alias{metagenomeAnnotation-class}
\title{metagenomeAnnotation-class object}
\value{
metagenomeAnnotation class object
}
\description{
Object contains taxonomic annotation and sequence data for an experiment .
The class extends the \link[Biobase]{AnnotatedDataFrame} class with a slot
for metadata including information on the database source and methods used to
perform the taxonomic assignment. Additionally, the user can include
experiment sequence data in the object for use in downstream analysis.
}
\section{Slots}{

\describe{
\item{\code{metadata}}{list}

\item{\code{referenceDbSeqData}}{DNAStringSet}

\item{\code{experimentTreeData}}{phylo}

\item{\code{experimentSeqData}}{DNAStringSet}
}}
\examples{
data(msd16s_metagenomeAnnotation)
}

