% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mgDb-class.R
\docType{class}
\name{MgDb-class}
\alias{MgDb}
\alias{MgDb-class}
\alias{mgdb}
\title{Metagenome Database class}
\usage{
# library(greengenes13.5MgDb)
}
\description{
The MgDb-class object contains sequence and taxonomic data for a 16S rRNA
taxonomic database, see the \pkg{greengenes13.5MgDb} package as an example
database.
}
\section{Fields}{

\describe{
\item{\code{taxa}}{taxonomic information for database sequences}

\item{\code{seq}}{database reference sequences}

\item{\code{metadata}}{associated metadata for the database}
}}
\note{
Currently the only database with a MgDb package is the
  \href{http://greengenes.secondgenome.com/}{Greengenes database} (version
  13.5), additional packages are planned.
}
\examples{
# see vignette
}

