% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mgDb-class.R
\docType{class}
\name{MgDb-class}
\alias{MgDb}
\alias{MgDb-class}
\alias{mgdb}
\alias{show,MgDb-method}
\title{Metagenome Database class}
\usage{
\S4method{show}{MgDb}(object)
}
\arguments{
\item{object}{MgDb-class object}
}
\description{
Metagenome Database class

Display summary of MgDb-class object
}
\section{Fields}{

\describe{
\item{\code{taxa}}{taxonomic information for database sequences}

\item{\code{seq}}{database reference sequences}

\item{\code{metadata}}{associated metadata for the database}
}}

