% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mgDb-class.R
\docType{class}
\name{MgDb-class}
\alias{MgDb}
\alias{MgDb-class}
\title{Metagenome Database class}
\description{
Metagenome Database class
}
\section{Fields}{

\describe{
\item{\code{taxa}}{taxonomic information for database sequences}

\item{\code{seq}}{database reference sequences}

\item{\code{metadata}}{associated metadata for the database}
}}

