% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb-class.R
\docType{class}
\name{MgDb-class}
\alias{MgDb-class}
\alias{mgdb}
\title{Metagenome Database class}
\value{
MgDb-class object
}
\description{
The MgDb-class object contains sequence, taxonomic data, and a phylogenetic
tree (optional) for a 16S rRNA taxonomic database, see the
\pkg{greengenes13.5MgDb} package as an example database. The
\code{get_gg13.8_85MgDb()} function in \pkg{metagenomeFeatures} exports a small
subset of the database in the \pkg{greengenes13.5MgDb} annotation package as
an example MgDb-class object.
}
\section{Slots}{

\describe{
\item{\code{seq}}{database reference sequences}

\item{\code{tree}}{reference phylogenetic tree}

\item{\code{taxa}}{database taxonomy}

\item{\code{metadata}}{associated metadata for the database}
}}

\note{
Currently the only database with a MgDb package is the
  \href{http://greengenes.secondgenome.com/}{Greengenes database} (version
  13.5), additional packages are planned.
}
\examples{
# example MgDb-class object, Greengenes 13.8 85\% OTUs database.
gg85 <- get_gg13.8_85MgDb()
}
