% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb-class.R
\docType{class}
\name{MgDb-class}
\alias{MgDb}
\alias{MgDb-class}
\alias{mgdb}
\title{Metagenome Database class}
\usage{
# library(greengenes13.5MgDb)
}
\description{
The MgDb-class object contains sequence and taxonomic data for a 16S rRNA
taxonomic database, see the \pkg{greengenes13.5MgDb} package as an example
database. The \code{get_demoMgDb} function exports a small subset of the database in \pkg{greengenes13.5MgDb}\pkg{metagenomeFeatures} package as an example of a MgDb-class object.
}
\section{Fields}{

\describe{
\item{\code{taxa}}{taxonomic information for database sequences}

\item{\code{seq}}{database reference sequences}

\item{\code{tree}}{reference phylogenetic tree}

\item{\code{taxa_file}}{name of sqlite db}

\item{\code{tree_file}}{name of phylogenetic tree file}

\item{\code{metadata}}{associated metadata for the database}
}}
\note{
Currently the only database with a MgDb package is the
  \href{http://greengenes.secondgenome.com/}{Greengenes database} (version
  13.5), additional packages are planned.
}
\examples{
# example MgDb-class object, a small subset of the Greengenes 13.5 database.
get_demoMgDb()
}

