% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb-class.R
\name{mgdb_meta}
\alias{mgdb_meta}
\title{MgDb metadata slot accessor}
\usage{
mgdb_meta(mgdb)
}
\arguments{
\item{mgdb}{MgDb class object}
}
\value{
list
}
\description{
MgDb metadata slot accessor
}
\examples{
# mgdb_meta(demoMgDb)
}

