% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb-class.R
\name{mgdb_tree}
\alias{mgdb_tree}
\title{MgDb tree slot accessor}
\usage{
mgdb_tree(mgdb)
}
\arguments{
\item{mgdb}{MgDb class object}
}
\value{
phylo class object
}
\description{
MgDb tree slot accessor
}
\examples{
demoMgDb <- get_demoMgDb()
mgdb_tree(demoMgDb)
}
