% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgFeatures-class.R
\name{mgF_tree}
\alias{mgF_tree}
\title{mgFeatures refDbTree slot accessor}
\usage{
mgF_tree(mgF)
}
\arguments{
\item{mgF}{mgFeatures class object}
}
\value{
phylo class object
}
\description{
mgFeatures refDbTree slot accessor
}
\examples{
data(mock_mgF)
mgF_tree(mock_mgF)
}

