% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgFeatures-class.R
\docType{class}
\name{mgFeatures-class}
\alias{mgFeatures-class}
\title{mgFeature-class object}
\value{
mgFeature class object
}
\description{
Object contains taxonomic annotation and reference sequence data for
classified OTUs. The class extends the \link[Biobase]{AnnotatedDataFrame}
class with a slot with a subset of the reference database sequences and
phylogenetic tree for taxonomically classified OTUs, along with an additional
slot for metadata including information on the database source.
}
\section{Slots}{

\describe{
\item{\code{metadata}}{list}

\item{\code{refDbSeq}}{DNAStringSet}

\item{\code{refDbTree}}{phyloOrNULL}
}}
\examples{
data(msd16s_mgFeatures)
}

