% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgFeatures-class.R
\name{mgFeatures}
\alias{mgFeatures}
\title{mgFeatures-class constructor}
\usage{
mgFeatures(taxa, tree = NULL, seq = NULL, metadata)
}
\arguments{
\item{taxa}{a DataFrame-class or object that can be coerced into a DataFrame}

\item{tree}{a phylo-class object with phylogenetic tree}

\item{seq}{DNAStringSet-object with feature sequences}

\item{metadata}{a list}
}
\value{
mgFeatures-class object
}
\description{
mgFeatures-class constructor
}
\examples{
mgFeatures(taxa = data.frame(), metadata = list())
}
