% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mgDb-class.R
\docType{methods}
\name{select}
\alias{select}
\alias{select,MgDb-method}
\title{Querying MgDb objects}
\usage{
select(mgdb, type, ...)

\S4method{select}{MgDb}(mgdb, type, keys = NULL, keytype = NULL,
  columns = "all")
}
\arguments{
\item{mgdb}{MgDb class object}

\item{type}{either "taxa", "seq", or "both". "taxa" and "seq" only queries
the taxonomy and sequences databases respectively. "both" queries both the
taxonomy and sequence database.}

\item{...}{additional arguments passed to select function}

\item{keys}{specific taxonomic groups to select for}

\item{keytype}{taxonomic level of keys}

\item{columns}{keytypes in taxonomy databse to return, all by default}
}
\value{
generates database, function does not return anything
}
\description{
Function for querying \code{\link{MgDb}} class objects, user defines the
taxonomic levels (\code{keytype}) and a vector of taxonomic names
(\code{keys}) being selected. If specific database ids are being selected for
use \code{keytype="Keys"}. Additionally, users can specify whether they want
only the taxonomic and sequence data, or both.
}
\examples{
\dontrun{
## need to install greengenes13.5MgDb from github
## https://github.com/HCBravoLab/greengenes13.5MgDb
## untill package is released on bioconductor
library(greengenes13.5MgDb)
# select taxa only
select(gg13.5MgDb, type = "taxa",
     keys = c("Vibrio", "Salmonella"),
     keytype = "Genus")

 # select seq only
select(gg13.5MgDb, type = "seq",
      keys = c("Vibrio", "Salmonella"),
      keytype = "Genus")

# select both taxa and seq
select(gg13.5MgDb, type = "both",
       keys = c("Vibrio", "Salmonella"),
       keytype = "Genus")
}
}

