% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mgDb-class.R
\docType{methods}
\name{select,MgDb-method}
\alias{select,MgDb-method}
\title{Function for querying MgDb class objects}
\usage{
\S4method{select}{MgDb}(mgdb, type, keys = NULL, keytype = NULL,
  ids = NULL, columns = "all")
}
\arguments{
\item{mgdb}{MgDb class object}

\item{type}{either "taxa", "seq", or "both". "taxa" and "seq" only queries the taxonomy and sequences databases respectively. "both" queries both the taxonomy and sequence database.}

\item{keys}{specific taxonomic groups to select for}

\item{keytype}{taxonomic level of keys}

\item{ids}{sequence ids to select}

\item{columns}{keytypes in taxonomy databse to return, all by default}
}
\value{
generates database, function does not return anything
}
\description{
Function for querying MgDb class objects
}

