% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb_method_select.R
\docType{methods}
\name{mgDb_select}
\alias{mgDb_select}
\alias{mgDb_select,MgDb-method}
\alias{select,MgDb-method}
\title{Querying MgDb objects}
\usage{
mgDb_select(mgdb, type, ...)

\S4method{mgDb_select}{MgDb}(mgdb, type, keys = NULL, keytype = NULL,
  columns = "all")
}
\arguments{
\item{mgdb}{MgDb class object}

\item{type}{either "taxa", "seq", "tree", "all" or a character vector of
types. "taxa", "seq", and "tree" only query the reference taxonomy,
sequences, and phylogenetic tree respectively. "all" queries the reference
taxonomy, sequence, and phylogenetic tree.}

\item{...}{additional arguments passed to select function}

\item{keys}{specific taxonomic groups to select for}

\item{keytype}{taxonomic level of keys}

\item{columns}{keytypes in taxonomy databse to return, all by default}
}
\value{
returned object depends on type: 'taxa' - dataframe with taxa
  information; 'seq' - DNAStringSet with seqeunce data; 'tree' - phylogenetic
  tree of class phylo; 'all' - list with the dataframe, DNAStringSet, and
  phylo.
}
\description{
Function for querying \code{\link{MgDb}} class objects, user defines the
taxonomic levels (\code{keytype}) and a vector of taxonomic names
(\code{keys}) being selected. If specific database ids are being selected for
use \code{keytype="Keys"}. Additionally, users can specify whether they want
only the taxonomic and sequence data, or both.
}
\examples{
demoMgDb <- get_demoMgDb()
# select taxa only
mgDb_select(demoMgDb, type = "taxa",
     keys = c("Vibrionaceae", "Enterobacteriaceae"),
     keytype = "Family")

# select seq only
 mgDb_select(demoMgDb, type = "seq",
      keys = c("Vibrionaceae", "Enterobacteriaceae"),
      keytype = "Family")

# select all taxa, seq, and tree
mgDb_select(demoMgDb, type = "all",
       keys = c("Vibrionaceae", "Enterobacteriaceae"),
       keytype = "Family")
}
