% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb-class.R
\docType{methods}
\name{select}
\alias{select}
\alias{select,MgDb-method}
\title{Querying MgDb objects}
\usage{
select(mgdb, type, ...)

\S4method{select}{MgDb}(mgdb, type, keys = NULL, keytype = NULL,
  columns = "all")
}
\arguments{
\item{mgdb}{MgDb class object}

\item{type}{either "taxa", "seq", "tree", "all" or a character vector of types. "taxa", "seq", and "tree" only query
the reference taxonomy, sequences, and phylogenetic tree respectively.
"all" queries the reference taxonomy, sequence, and phylogenetic tree.}

\item{...}{additional arguments passed to select function}

\item{keys}{specific taxonomic groups to select for}

\item{keytype}{taxonomic level of keys}

\item{columns}{keytypes in taxonomy databse to return, all by default}
}
\value{
returned object depends on type, for 'taxa' -dataframe with taxa
  information, 'seq' a DNAStringSet with seqeunce data, 'tree' a phylogenetic tree of class phylo, 'all' a list with
  the dataframe, DNAStringSet, and phylo.
}
\description{
Function for querying \code{\link{MgDb}} class objects, user defines the
taxonomic levels (\code{keytype}) and a vector of taxonomic names
(\code{keys}) being selected. If specific database ids are being selected for
use \code{keytype="Keys"}. Additionally, users can specify whether they want
only the taxonomic and sequence data, or both.
}
\examples{
demoMgDb <- get_demoMgDb()
# select taxa only
select(demoMgDb, type = "taxa",
     keys = c("Vibrio", "Salmonella"),
     keytype = "Genus")

# select seq only
 select(demoMgDb, type = "seq",
      keys = c("Vibrio", "Salmonella"),
      keytype = "Genus")

# select all taxa, seq, and tree
select(demoMgDb, type = "all",
       keys = c("Vibrio", "Salmonella"),
       keytype = "Genus")
}

