% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb-class.R
\docType{methods}
\name{show,MgDb-method}
\alias{show,MgDb-method}
\title{Display summary of MgDb-class object}
\usage{
\S4method{show}{MgDb}(object)
}
\arguments{
\item{object}{MgDb-class object}
}
\value{
MgDb-class summary
}
\description{
Display summary of MgDb-class object
}
