% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metagenomeAnnotation-class.R
\name{split_by}
\alias{split_by}
\title{Split metagenomeAnnotation Object by Taxa}
\usage{
split_by(object, taxa_level)
}
\arguments{
\item{object}{}

\item{taxa_level}{}
}
\value{
list of metagenomeAnnotation objects
}
\description{
Split metagenomeAnnotation Object by Taxa
}
\examples{
split_by(mgAnno, "Phylum")
}

