% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomeAnnotation-class.R
\name{split_by}
\alias{split_by}
\title{Split metagenomeAnnotation Object by Taxa}
\usage{
split_by(mgAnno, taxa_level)
}
\arguments{
\item{mgAnno}{metagenomeAnnotation class object}

\item{taxa_level}{taxonomic level used to split the metagenomeAnnotation
object at}
}
\value{
list of metagenomeAnnotation objects
}
\description{
Function takes a \link{metagenomeAnnotation-class}
object and splits it based on a user provided taxonomic level, returning a
list of \link{metagenomeAnnotation-class} objects for
each unique taxa at that level.
}
\examples{
data(msd16s_metagenomeAnnotation)
split_by(msd16s_metagenomeAnnotation, "Phylum")
}

