% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb-taxa_accessors.R
\docType{methods}
\name{taxa_columns}
\alias{taxa_columns}
\alias{taxa_columns,MgDb-method}
\title{Column names for MgDb taxonomy slot object}
\usage{
taxa_columns(mgdb)

\S4method{taxa_columns}{MgDb}(mgdb)
}
\arguments{
\item{mgdb}{object of MgDB class}
}
\value{
character vector
}
\description{
Column names for MgDb taxonomy slot object
}
\note{
Same function as \code{\link{taxa_keytypes}}.
}
\examples{
demoMgDb <- get_demoMgDb()
taxa_columns(demoMgDb)
}
