% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mgDb-taxa_accessors.R
\docType{methods}
\name{taxa_keys}
\alias{taxa_keys}
\alias{taxa_keys,MgDb-method}
\title{Taxonomy values for a given keytype}
\usage{
taxa_keys(mgdb, keytype)

\S4method{taxa_keys}{MgDb}(mgdb, keytype)
}
\arguments{
\item{mgdb}{object of MgDb class}

\item{keytype}{taxonomic classification level}
}
\value{
tbl_df
}
\description{
Taxonomy values for a given keytype
}
\section{Methods (by class)}{
\itemize{
\item \code{MgDb}: 
}}
\examples{
library(greengenes13.5MgDb)
taxa_keys(gg13.5MgDb, keytype = "Phylum")
}
\seealso{
Other MgDb_accessors: \code{\link{taxa_columns}},
  \code{\link{taxa_columns,MgDb-method}};
  \code{\link{taxa_keytypes}},
  \code{\link{taxa_keytypes,MgDb-method}}
}

