% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mgDb-taxa_accessors.R
\docType{methods}
\name{taxa_columns}
\alias{taxa_columns}
\alias{taxa_columns,MgDb-method}
\alias{taxa_keytypes}
\alias{taxa_keytypes,MgDb-method}
\title{Column names for MgDb taxonomy slot object}
\usage{
taxa_columns(mgdb)

\S4method{taxa_columns}{MgDb}(mgdb)

taxa_keytypes(mgdb)

\S4method{taxa_keytypes}{MgDb}(mgdb)
}
\arguments{
\item{mgdb}{object of MgDB class}
}
\value{
tbl_df
}
\description{
Column names for MgDb taxonomy slot object
}
\examples{
\dontrun{
# need to install greengenes13.5MgDb from github https://github.com/HCBravoLab/greengenes13.5MgDb untill package is released on bioconductor
library(greengenes13.5MgDb)
taxa_keytypes(gg13.5MgDb)
}
}

