% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgDb-taxa_accessors.R
\docType{methods}
\name{taxa_keytypes}
\alias{taxa_keytypes}
\alias{taxa_keytypes,MgDb-method}
\title{Column names for MgDb taxonomy slot object}
\usage{
taxa_keytypes(mgdb)

\S4method{taxa_keytypes}{MgDb}(mgdb)
}
\arguments{
\item{mgdb}{object of MgDB class}
}
\value{
tbl_df
}
\description{
Column names for MgDb taxonomy slot object
}
\examples{
demoMgDb <- get_demoMgDb()
taxa_keytypes(demoMgDb)
}
